refactor: useDeleteConnectionMutation
